# Dont calculate buildpack layer diffIDs twice when creating a builder